Update to add labels to graph-view
The original implementation used ids for
labels, however one can add labels to
both vertices and edges in digraph.

This patch add vertex labels as title which
shows the content as you hover over the vertex.

The labels are used on the edge instead of its
id which is rather cryptic.

Added small example graph to show the intent.